修复登录新浪微博闪退的问题 改善播放页手势用户体验 自动读取剪贴板的磁力链内容 方便下载 加快本地视频缩略图获取速度 添加设置同屏弹幕数量功能 修改本地文件列表UI 增加账号登录系统

+++ DanDanPlayForiOS/Widget/JHMediaPlayer/JHMediaPlayer.h
@@ -36,11 +36,11 @@ typedef void(^SnapshotCompleteBlock)(UIImage *image, NSError *error);
  @param totalSeconds 秒数
  @return 指定格式
  */
-CG_INLINE NSString *jh_mediaFormatterTime(int totalSeconds) {
-    int seconds = totalSeconds % 60;
-    int minutes = (totalSeconds - seconds) / 60;
+CG_INLINE NSString *jh_mediaFormatterTime(NSInteger totalSeconds) {
+    NSInteger seconds = totalSeconds % 60;
+    NSInteger minutes = (totalSeconds - seconds) / 60;
     
-    return [NSString stringWithFormat:@"%.2d:%.2d", minutes, seconds];
+    return [NSString stringWithFormat:@"%.2ld:%.2ld", (long)minutes, seconds];
 }
 
 @class JHMediaPlayer;
